# Port `equal` from THC to ATen (CUDA)