--- a/aten/src/THC/generic/THCTensorMathPairwise.cu
+++ b/aten/src/THC/generic/THCTensorMathPairwise.cu
@@ -4,41 +4,6 @@
 
 #include <ATen/NamedTensorUtils.h>
 
-static int THCTensor_(equalImpl)(THCState *state, THCTensor *self_, THCTensor *src_)
-{
-  THCAssertSameGPU(THCTensor_(checkGPU)(state, 2, self_, src_));
-  if (!THCTensor_(isSameSizeAs(state, self_, src_))) {
-    return 0;
-  }
-
-  // This is not as efficient as TH, but the basic idea: create a buffer that stores
-  // 1 if the two tensors are equal at a position, otherwise 0. If the minimum value
-  // in this buffer is 1, the two tensors are equal, otherwise they are not
-
-  // Both tensors are empty
-  if(THTensor_(nElement)(self_) == 0) return true;
-
-  THCudaByteTensor *buf = at::empty_like(THTensor_wrap(self_), at::kByte).unsafeReleaseTensorImpl();
-
-  if (!THC_pointwiseApply3<uint8_t, scalar_t, scalar_t>(state, buf, self_, src_, TensorEQOp<scalar_t, unsigned char>())) {
-    THArgCheck(false, 2, CUTORCH_DIM_WARNING);
-  }
-
-  unsigned char min = THTensor_wrap(buf).min().item<unsigned char>();
-
-  THCudaByteTensor_free(state, buf);
-
-  return min != 0;
-}
-
-int THCTensor_(equal)(THCState *state, THCTensor *self_, THCTensor *src_) {
-  if (!at::namedinference::are_names_equal(self_, src_)) {
-    return 0;
-  }
-  at::NoNamesGuard guard;
-  return THCTensor_(equalImpl)(state, self_, src_);
-}
-
 #if !defined(THC_REAL_IS_BOOL)
 
 void THCTensor_(mul)(THCState *state, THCTensor *self_, THCTensor *src_, scalar_t value)