@Baranowski 2020-04-13
I don't see anything for `equal` in the `benchmarks/` folder. Any suggestions on what usage patterns I should benchmark for?