@Baranowski 2020-04-21
@VitalyFedyunin should I benchmark this change?