@VitalyFedyunin 2020-04-21
Hi! Yes, simple timeit benchmark with different tensor shapes(sizes) will suffice.