@Baranowski 2020-04-22
I wasn't expecting this in such a simple change but I'm seeing a perf regression. I will need to dig into this.