@Baranowski 2020-07-03
@VitalyFedyunin I have rebased again. This is ready for merge