@Baranowski 2020-07-06
Just a heads up: I won't have the time to work on this after two weeks from now. I think this should be ready to merge.